Avoid breaking change and move non-schema constants back to `arrow-data`
